Reinstate "Replace deprecated method for initializing nibs (10.8)"
This reinstates commit 6ce0d06 ("Replace deprecated method for
initializing nibs (10.8)"). The problem was the use of NSWorkspace's
`isFilePackageAtPath:` to determine if the nib was compiled or not,
since it appears that Xcode.app defines .nib as a file package and not
the OS. Therefore, the previous check would fail for users without
Xcode installed.

This version just simply checks if the file keyedobjects.nib exist.